+++ cucumber-tsflow/src/types/global.d.ts
@@ -9,6 +9,12 @@ declare global {
 	var experimentalDecorators: boolean;
 	// eslint-disable-next-line no-var
 	var __LOADER_WORKER: boolean;
+	/**
+	 * Source maps of the modules the esbuild ESM loader transpiled on this thread, by module URL, set in
+	 * `transpilers/esm/loader-utils.mjs` and read by `utils/our-callsite.ts`.
+	 */
+	// eslint-disable-next-line no-var
+	var __CUCUMBER_TSFLOW_SOURCE_MAPS: Map<string, string> | undefined;
 }
 
 export {};